+++ internal/retry/retry_test.go
@@ -22,7 +22,6 @@ package retry
 
 import (
 	"bytes"
-	"context"
 	"testing"
 	"time"
 
@@ -38,17 +37,19 @@ func TestMiddleware(t *testing.T) {
 	type testStruct struct {
 		msg string
 
-		retries      uint
-		retryTimeout time.Duration
-		retryBackoff backoff.Strategy
+		policyProvider *procedurePolicyProvider
 
 		actions []MiddlewareAction
 	}
 	tests := []testStruct{
 		{
-			msg:          "no retry",
-			retries:      1,
-			retryTimeout: time.Millisecond * 500,
+			msg: "no retry",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(1),
+					MaxRequestTimeout(testtime.Millisecond*500),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				RequestAction{
 					request: &transport.Request{
@@ -56,7 +57,7 @@ func TestMiddleware(t *testing.T) {
 						Procedure: "proc",
 						Body:      bytes.NewBufferString("body"),
 					},
-					reqTimeout: time.Second,
+					reqTimeout: testtime.Second,
 					events: []*OutboundEvent{
 						{
 							WantService:   "serv",
@@ -70,9 +71,13 @@ func TestMiddleware(t *testing.T) {
 			},
 		},
 		{
-			msg:          "single retry",
-			retries:      1,
-			retryTimeout: time.Millisecond * 500,
+			msg: "single retry",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(1),
+					MaxRequestTimeout(testtime.Millisecond*500),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				RequestAction{
 					request: &transport.Request{
@@ -80,7 +85,7 @@ func TestMiddleware(t *testing.T) {
 						Procedure: "proc",
 						Body:      bytes.NewBufferString("body"),
 					},
-					reqTimeout: time.Second,
+					reqTimeout: testtime.Second,
 					events: []*OutboundEvent{
 						{
 							WantService:   "serv",
@@ -100,9 +105,13 @@ func TestMiddleware(t *testing.T) {
 			},
 		},
 		{
-			msg:          "multiple retries",
-			retries:      4,
-			retryTimeout: time.Millisecond * 500,
+			msg: "multiple retries",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(4),
+					MaxRequestTimeout(testtime.Millisecond*500),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				RequestAction{
 					request: &transport.Request{
@@ -110,7 +119,7 @@ func TestMiddleware(t *testing.T) {
 						Procedure: "proc",
 						Body:      bytes.NewBufferString("body"),
 					},
-					reqTimeout: time.Second,
+					reqTimeout: testtime.Second,
 					events: []*OutboundEvent{
 						{
 							WantService:   "serv",
@@ -122,7 +131,7 @@ func TestMiddleware(t *testing.T) {
 							WantService:   "serv",
 							WantProcedure: "proc",
 							WantBody:      "body",
-							GiveError:     yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", time.Millisecond*300),
+							GiveError:     yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*300),
 						},
 						{
 							WantService:   "serv",
@@ -148,9 +157,13 @@ func TestMiddleware(t *testing.T) {
 			},
 		},
 		{
-			msg:          "immediate hard failure",
-			retries:      1,
-			retryTimeout: time.Millisecond * 500,
+			msg: "immediate hard failure",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(1),
+					MaxRequestTimeout(testtime.Millisecond*500),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				RequestAction{
 					request: &transport.Request{
@@ -158,7 +171,7 @@ func TestMiddleware(t *testing.T) {
 						Procedure: "proc",
 						Body:      bytes.NewBufferString("body"),
 					},
-					reqTimeout: time.Second,
+					reqTimeout: testtime.Second,
 					events: []*OutboundEvent{
 						{
 							WantService:   "serv",
@@ -172,9 +185,13 @@ func TestMiddleware(t *testing.T) {
 			},
 		},
 		{
-			msg:          "retry once, then hard failure",
-			retries:      1,
-			retryTimeout: time.Millisecond * 500,
+			msg: "retry once, then hard failure",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(1),
+					MaxRequestTimeout(testtime.Millisecond*500),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				RequestAction{
 					request: &transport.Request{
@@ -182,7 +199,7 @@ func TestMiddleware(t *testing.T) {
 						Procedure: "proc",
 						Body:      bytes.NewBufferString("body"),
 					},
-					reqTimeout: time.Second,
+					reqTimeout: testtime.Second,
 					events: []*OutboundEvent{
 						{
 							WantService:   "serv",
@@ -202,9 +219,13 @@ func TestMiddleware(t *testing.T) {
 			},
 		},
 		{
-			msg:          "ctx timeout less than retry timeout",
-			retries:      1,
-			retryTimeout: time.Millisecond * 500,
+			msg: "ctx timeout less than retry timeout",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(1),
+					MaxRequestTimeout(testtime.Millisecond*500),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				RequestAction{
 					request: &transport.Request{
@@ -212,9 +233,10 @@ func TestMiddleware(t *testing.T) {
 						Procedure: "proc",
 						Body:      bytes.NewBufferString("body"),
 					},
-					reqTimeout: time.Millisecond * 300,
+					reqTimeout: testtime.Millisecond * 300,
 					events: []*OutboundEvent{
 						{
+							WantTimeout:   testtime.Millisecond * 300,
 							WantService:   "serv",
 							WantProcedure: "proc",
 							WantBody:      "body",
@@ -226,9 +248,13 @@ func TestMiddleware(t *testing.T) {
 			},
 		},
 		{
-			msg:          "ctx timeout less than retry timeout",
-			retries:      1,
-			retryTimeout: time.Millisecond * 50,
+			msg: "ctx timeout less than retry timeout",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(1),
+					MaxRequestTimeout(testtime.Millisecond*50),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				RequestAction{
 					request: &transport.Request{
@@ -236,16 +262,18 @@ func TestMiddleware(t *testing.T) {
 						Procedure: "proc",
 						Body:      bytes.NewBufferString("body"),
 					},
-					reqTimeout: time.Millisecond * 75,
+					reqTimeout: testtime.Millisecond * 75,
 					events: []*OutboundEvent{
 						{
+							WantTimeout:    testtime.Millisecond * 50,
 							WantService:    "serv",
 							WantProcedure:  "proc",
 							WantBody:       "body",
 							WaitForTimeout: true,
-							GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", time.Millisecond*50),
+							GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*50),
 						},
 						{
+							WantTimeout:   testtime.Millisecond * 25,
 							WantService:   "serv",
 							WantProcedure: "proc",
 							WantBody:      "body",
@@ -257,9 +285,13 @@ func TestMiddleware(t *testing.T) {
 			},
 		},
 		{
-			msg:          "no ctx timeout",
-			retries:      1,
-			retryTimeout: time.Millisecond * 50,
+			msg: "no ctx timeout",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(1),
+					MaxRequestTimeout(testtime.Millisecond*50),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				RequestAction{
 					request: &transport.Request{
@@ -269,13 +301,15 @@ func TestMiddleware(t *testing.T) {
 					},
 					events: []*OutboundEvent{
 						{
+							WantTimeout:    testtime.Millisecond * 50,
 							WantService:    "serv",
 							WantProcedure:  "proc",
 							WantBody:       "body",
 							WaitForTimeout: true,
-							GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", time.Millisecond*50),
+							GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*50),
 						},
 						{
+							WantTimeout:   testtime.Millisecond * 50,
 							WantService:   "serv",
 							WantProcedure: "proc",
 							WantBody:      "body",
@@ -287,9 +321,13 @@ func TestMiddleware(t *testing.T) {
 			},
 		},
 		{
-			msg:          "exhaust retries",
-			retries:      1,
-			retryTimeout: time.Millisecond * 50,
+			msg: "exhaust retries",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(1),
+					MaxRequestTimeout(testtime.Millisecond*50),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				RequestAction{
 					request: &transport.Request{
@@ -297,15 +335,17 @@ func TestMiddleware(t *testing.T) {
 						Procedure: "proc",
 						Body:      bytes.NewBufferString("body"),
 					},
-					reqTimeout: time.Millisecond * 400,
+					reqTimeout: testtime.Millisecond * 400,
 					events: []*OutboundEvent{
 						{
+							WantTimeout:   testtime.Millisecond * 50,
 							WantService:   "serv",
 							WantProcedure: "proc",
 							WantBody:      "body",
 							GiveError:     yarpcerrors.InternalErrorf("unexpected error 1"),
 						},
 						{
+							WantTimeout:   testtime.Millisecond * 50,
 							WantService:   "serv",
 							WantProcedure: "proc",
 							WantBody:      "body",
@@ -317,9 +357,13 @@ func TestMiddleware(t *testing.T) {
 			},
 		},
 		{
-			msg:          "Reset Error",
-			retries:      1,
-			retryTimeout: time.Millisecond * 50,
+			msg: "Reset Error",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(1),
+					MaxRequestTimeout(testtime.Millisecond*50),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				RequestAction{
 					request: &transport.Request{
@@ -327,9 +371,10 @@ func TestMiddleware(t *testing.T) {
 						Procedure: "proc",
 						Body:      bytes.NewBufferString("body"),
 					},
-					reqTimeout: time.Millisecond * 400,
+					reqTimeout: testtime.Millisecond * 400,
 					events: []*OutboundEvent{
 						{
+							WantTimeout:   testtime.Millisecond * 50,
 							WantService:   "serv",
 							WantProcedure: "proc",
 							// We have explicitly not read the body, which will not exhaust the
@@ -342,10 +387,14 @@ func TestMiddleware(t *testing.T) {
 			},
 		},
 		{
-			msg:          "backoff timeout",
-			retries:      1,
-			retryTimeout: time.Millisecond * 50,
-			retryBackoff: newFixedBackoff(time.Millisecond * 25),
+			msg: "backoff timeout",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(1),
+					MaxRequestTimeout(testtime.Millisecond*50),
+					BackoffStrategy(newFixedBackoff(testtime.Millisecond*25)),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				RequestAction{
 					request: &transport.Request{
@@ -353,16 +402,18 @@ func TestMiddleware(t *testing.T) {
 						Procedure: "proc",
 						Body:      bytes.NewBufferString("body"),
 					},
-					reqTimeout: time.Millisecond * 100,
+					reqTimeout: testtime.Millisecond * 100,
 					events: []*OutboundEvent{
 						{
+							WantTimeout:    testtime.Millisecond * 50,
 							WantService:    "serv",
 							WantProcedure:  "proc",
 							WantBody:       "body",
 							WaitForTimeout: true,
-							GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", time.Millisecond*50),
+							GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*50),
 						},
 						{
+							WantTimeout:   testtime.Millisecond * 25,
 							WantService:   "serv",
 							WantProcedure: "proc",
 							WantBody:      "body",
@@ -374,10 +425,14 @@ func TestMiddleware(t *testing.T) {
 			},
 		},
 		{
-			msg:          "sequential backoff timeout",
-			retries:      2,
-			retryTimeout: time.Millisecond * 100,
-			retryBackoff: newSequentialBackoff(time.Millisecond * 50),
+			msg: "sequential backoff timeout",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(2),
+					MaxRequestTimeout(testtime.Millisecond*100),
+					BackoffStrategy(newSequentialBackoff(testtime.Millisecond*50)),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				RequestAction{
 					request: &transport.Request{
@@ -385,27 +440,33 @@ func TestMiddleware(t *testing.T) {
 						Procedure: "proc",
 						Body:      bytes.NewBufferString("body"),
 					},
-					reqTimeout: time.Millisecond * 400,
+					reqTimeout: testtime.Millisecond * 400,
 					events: []*OutboundEvent{
 						{
-							WantService:    "serv",
-							WantProcedure:  "proc",
-							WantBody:       "body",
-							WaitForTimeout: true,
-							GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", time.Millisecond*50),
+							WantTimeout:       testtime.Millisecond * 100,
+							WantTimeoutBounds: testtime.Millisecond * 20,
+							WantService:       "serv",
+							WantProcedure:     "proc",
+							WantBody:          "body",
+							WaitForTimeout:    true,
+							GiveError:         yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*50),
 						},
 						{
-							WantService:    "serv",
-							WantProcedure:  "proc",
-							WantBody:       "body",
-							WaitForTimeout: true,
-							GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", time.Millisecond*50),
+							WantTimeout:       testtime.Millisecond * 100,
+							WantTimeoutBounds: testtime.Millisecond * 20,
+							WantService:       "serv",
+							WantProcedure:     "proc",
+							WantBody:          "body",
+							WaitForTimeout:    true,
+							GiveError:         yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*50),
 						},
 						{
-							WantService:   "serv",
-							WantProcedure: "proc",
-							WantBody:      "body",
-							GiveRespBody:  "respbody",
+							WantTimeout:       testtime.Millisecond * 50,
+							WantTimeoutBounds: testtime.Millisecond * 20,
+							WantService:       "serv",
+							WantProcedure:     "proc",
+							WantBody:          "body",
+							GiveRespBody:      "respbody",
 						},
 					},
 					wantBody: "respbody",
@@ -413,10 +474,14 @@ func TestMiddleware(t *testing.T) {
 			},
 		},
 		{
-			msg:          "backoff context will timeout",
-			retries:      2,
-			retryTimeout: time.Millisecond * 30,
-			retryBackoff: newFixedBackoff(time.Millisecond * 5000),
+			msg: "backoff context will timeout",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(2),
+					MaxRequestTimeout(testtime.Millisecond*30),
+					BackoffStrategy(newFixedBackoff(testtime.Millisecond*5000)),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				RequestAction{
 					request: &transport.Request{
@@ -424,25 +489,32 @@ func TestMiddleware(t *testing.T) {
 						Procedure: "proc",
 						Body:      bytes.NewBufferString("body"),
 					},
-					reqTimeout: time.Millisecond * 60,
+					reqTimeout: testtime.Millisecond * 60,
 					events: []*OutboundEvent{
 						{
-							WantService:    "serv",
-							WantProcedure:  "proc",
-							WantBody:       "body",
-							WaitForTimeout: true,
-							GiveError:      yarpcerrors.InternalErrorf("unexpected error 2"),
+							WantTimeout:       testtime.Millisecond * 30,
+							WantTimeoutBounds: testtime.Millisecond * 10,
+							WantService:       "serv",
+							WantProcedure:     "proc",
+							WantBody:          "body",
+							WaitForTimeout:    true,
+							GiveError:         yarpcerrors.InternalErrorf("unexpected error 2"),
 						},
 					},
-					wantError: yarpcerrors.InternalErrorf("unexpected error 2").Error(),
+					wantTimeLimit: testtime.Millisecond * 40,
+					wantError:     yarpcerrors.InternalErrorf("unexpected error 2").Error(),
 				},
 			},
 		},
 		{
-			msg:          "concurrent retries",
-			retries:      2,
-			retryTimeout: time.Millisecond * 50,
-			retryBackoff: newFixedBackoff(time.Millisecond * 25),
+			msg: "concurrent retries",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(2),
+					MaxRequestTimeout(testtime.Millisecond*50),
+					BackoffStrategy(newFixedBackoff(testtime.Millisecond*25)),
+				),
+			).provider,
 			actions: []MiddlewareAction{
 				ConcurrentAction{
 					Actions: []MiddlewareAction{
@@ -452,16 +524,18 @@ func TestMiddleware(t *testing.T) {
 								Procedure: "proc",
 								Body:      bytes.NewBufferString("body"),
 							},
-							reqTimeout: time.Millisecond * 100,
+							reqTimeout: testtime.Millisecond * 100,
 							events: []*OutboundEvent{
 								{
+									WantTimeout:    testtime.Millisecond * 50,
 									WantService:    "serv",
 									WantProcedure:  "proc",
 									WantBody:       "body",
 									WaitForTimeout: true,
-									GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", time.Millisecond*50),
+									GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*50),
 								},
 								{
+									WantTimeout:   testtime.Millisecond * 25,
 									WantService:   "serv",
 									WantProcedure: "proc",
 									WantBody:      "body",
@@ -476,9 +550,10 @@ func TestMiddleware(t *testing.T) {
 								Procedure: "proc2",
 								Body:      bytes.NewBufferString("body2"),
 							},
-							reqTimeout: time.Second,
+							reqTimeout: testtime.Second,
 							events: []*OutboundEvent{
 								{
+									WantTimeout:   testtime.Millisecond * 50,
 									WantService:   "serv2",
 									WantProcedure: "proc2",
 									WantBody:      "body2",
@@ -493,25 +568,142 @@ func TestMiddleware(t *testing.T) {
 								Procedure: "proc3",
 								Body:      bytes.NewBufferString("body3"),
 							},
-							reqTimeout: time.Millisecond * 100,
+							reqTimeout: testtime.Millisecond * 100,
 							events: []*OutboundEvent{
 								{
+									WantTimeout:    testtime.Millisecond * 50,
 									WantService:    "serv3",
 									WantProcedure:  "proc3",
 									WantBody:       "body3",
 									WaitForTimeout: true,
-									GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv3 procedure:proc3 ttl:%v", time.Millisecond*50),
+									GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv3 procedure:proc3 ttl:%v", testtime.Millisecond*50),
 								},
 								{
+									WantTimeout:    testtime.Millisecond * 25,
 									WantService:    "serv3",
 									WantProcedure:  "proc3",
 									WantBody:       "body3",
 									GiveRespBody:   "respbody",
 									WaitForTimeout: true,
-									GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv3 procedure:proc3 ttl:%v", time.Millisecond*25),
+									GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv3 procedure:proc3 ttl:%v", testtime.Millisecond*25),
 								},
 							},
-							wantError: yarpcerrors.DeadlineExceededErrorf("service:serv3 procedure:proc3 ttl:%v", time.Millisecond*25).Error(),
+							wantError: yarpcerrors.DeadlineExceededErrorf("service:serv3 procedure:proc3 ttl:%v", testtime.Millisecond*25).Error(),
+						},
+					},
+				},
+			},
+		},
+		{
+			msg: "multiple retry policies",
+			policyProvider: newPolicyProviderBuilder().registerDefault(
+				NewPolicy(
+					Retries(2),
+					MaxRequestTimeout(testtime.Millisecond*20),
+					BackoffStrategy(newFixedBackoff(testtime.Millisecond*25)),
+				),
+			).registerService(
+				"serviceRoute",
+				NewPolicy(
+					Retries(1),
+					MaxRequestTimeout(testtime.Millisecond*50),
+					BackoffStrategy(newFixedBackoff(testtime.Millisecond*50)),
+				),
+			).registerServiceProcedure(
+				"serviceRoute",
+				"procedureRoute",
+				NewPolicy(
+					Retries(0),
+					MaxRequestTimeout(testtime.Millisecond*100),
+				),
+			).registerServiceProcedure(
+				"serviceRoute",
+				"fakeProcedureRoute",
+				NewPolicy(
+					Retries(100),
+					MaxRequestTimeout(testtime.Millisecond*10000),
+				),
+			).provider,
+			actions: []MiddlewareAction{
+				ConcurrentAction{
+					Actions: []MiddlewareAction{
+						RequestAction{
+							request: &transport.Request{
+								Service:   "nonServiceRoute",
+								Procedure: "nonProcedureRoute",
+								Body:      bytes.NewBufferString("body1"),
+							},
+							reqTimeout: testtime.Millisecond * 200,
+							events: []*OutboundEvent{
+								{
+									WantTimeout:    testtime.Millisecond * 20,
+									WantService:    "nonServiceRoute",
+									WantProcedure:  "nonProcedureRoute",
+									WantBody:       "body1",
+									WaitForTimeout: true,
+									GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*20),
+								},
+								{
+									WantTimeout:    testtime.Millisecond * 20,
+									WantService:    "nonServiceRoute",
+									WantProcedure:  "nonProcedureRoute",
+									WantBody:       "body1",
+									WaitForTimeout: true,
+									GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*20),
+								},
+								{
+									WantTimeout:   testtime.Millisecond * 20,
+									WantService:   "nonServiceRoute",
+									WantProcedure: "nonProcedureRoute",
+									WantBody:      "body1",
+									GiveError:     yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*20),
+								},
+							},
+							wantError: yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*20).Error(),
+						},
+						RequestAction{
+							request: &transport.Request{
+								Service:   "serviceRoute",
+								Procedure: "nonProcedureRoute",
+								Body:      bytes.NewBufferString("body2"),
+							},
+							reqTimeout: testtime.Millisecond * 200,
+							events: []*OutboundEvent{
+								{
+									WantTimeout:    testtime.Millisecond * 50,
+									WantService:    "serviceRoute",
+									WantProcedure:  "nonProcedureRoute",
+									WantBody:       "body2",
+									WaitForTimeout: true,
+									GiveError:      yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*50),
+								},
+								{
+									WantTimeout:   testtime.Millisecond * 50,
+									WantService:   "serviceRoute",
+									WantProcedure: "nonProcedureRoute",
+									WantBody:      "body2",
+									GiveError:     yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*50),
+								},
+							},
+							wantError: yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*50).Error(),
+						},
+						RequestAction{
+							request: &transport.Request{
+								Service:   "serviceRoute",
+								Procedure: "procedureRoute",
+								Body:      bytes.NewBufferString("body3"),
+							},
+							reqTimeout: testtime.Millisecond * 200,
+							events: []*OutboundEvent{
+								{
+									WantTimeout:   testtime.Millisecond * 100,
+									WantService:   "serviceRoute",
+									WantProcedure: "procedureRoute",
+									WantBody:      "body3",
+									GiveError:     yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*100),
+								},
+							},
+							wantError: yarpcerrors.DeadlineExceededErrorf("service:serv procedure:proc ttl:%v", testtime.Millisecond*100).Error(),
 						},
 					},
 				},
@@ -522,15 +714,7 @@ func TestMiddleware(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.msg, func(t *testing.T) {
 			retry := NewUnaryMiddleware(
-				WithPolicyProvider(
-					func(context.Context, *transport.Request) *Policy {
-						return NewPolicy(
-							Retries(tt.retries),
-							MaxRequestTimeout(testtime.Scale(tt.retryTimeout)),
-							BackoffStrategy(tt.retryBackoff),
-						)
-					},
-				),
+				WithPolicyProvider(tt.policyProvider.GetPolicy),
 			)
 
 			ApplyMiddlewareActions(t, retry, tt.actions)
@@ -572,4 +756,29 @@ func (f *fixedBackoff) Backoff() backoff.Backoff {
 
 func (f *fixedBackoff) Duration(_ uint) time.Duration {
 	return f.boff
+}
+
+type policyProviderBuilder struct {
+	provider *procedurePolicyProvider
+}
+
+func newPolicyProviderBuilder() *policyProviderBuilder {
+	return &policyProviderBuilder{
+		provider: newProcedurePolicyProvider(),
+	}
+}
+
+func (pb *policyProviderBuilder) registerServiceProcedure(service, procedure string, pol *Policy) *policyProviderBuilder {
+	pb.provider.registerServiceProcedure(service, procedure, pol)
+	return pb
+}
+
+func (pb *policyProviderBuilder) registerService(service string, pol *Policy) *policyProviderBuilder {
+	pb.provider.registerService(service, pol)
+	return pb
+}
+
+func (pb *policyProviderBuilder) registerDefault(pol *Policy) *policyProviderBuilder {
+	pb.provider.registerDefault(pol)
+	return pb
 }